requirements.txt
httpx[http2]>=0.28.0
lxml>=5.0
beautifulsoup4>=4.12
pydantic>=2.0
playwright>=1.40
playwright-stealth>=1.0.6
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
python-dotenv>=1.0
tqdm>=4.66
tenacity>=8.2
anthropic>=0.40
fake-useragent>=1.5
